move util functions to util.js
